fix(hydrology): move random helpers to random header
