+++ HighMap/src/hydrology/flow_accumulation_stochastic.cpp
@@ -13,26 +13,13 @@
 #include "highmap/array.hpp"
 #include "highmap/gradient.hpp"
 #include "highmap/hydrology/hydrology.hpp"
+#include "highmap/random.hpp"
 
 namespace hmap
 {
 
 namespace
 {
-
-// SplitMix64 — deterministic per-sample uniforms
-uint64_t splitmix64(uint64_t x)
-{
-  x += 0x9e3779b97f4a7c15ULL;
-  x = (x ^ (x >> 30)) * 0xbf58476d1ce4e5b9ULL;
-  x = (x ^ (x >> 27)) * 0x94d049bb133111ebULL;
-  return x ^ (x >> 31);
-}
-
-float uniform01(uint64_t h)
-{
-  return (float)(h >> 40) / 16777216.f; // 24-bit mantissa in [0, 1)
-}
 
 float bilinear(const Array &a, float x, float y)
 {
